gpu_server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
SQLAlchemy==2.0.4
psycopg2-binary==2.9.5
pandas==1.5.3
numpy==1.23.5
gunicorn==20.1.0
requests==2.28.2
pydantic==1.10.5
python-dotenv==1.0.0